#90 Merge tests
